rootless: move GetAvailableGids to the rootless pkg
